fixed typecast warning

+++ src/ESPTelnet.cpp
@@ -149,7 +149,7 @@ size_t ESPTelnet::printf(const char* format, ...) {
       va_end(arg);
       return 0;
     };
-    if (len >= sizeof(loc_buf)) {
+    if (len >= (int)sizeof(loc_buf)) {
       temp = (char*)malloc(len + 1);
       if (temp == NULL) {
         va_end(arg);
